requirements.txt
fastapi
uvicorn[standard]
anthropic
python-dotenv
pydantic
google-cloud-firestore
